+++ src/main/scala/co.blocke.scalajack/json/reading/Helpers.scala
@@ -246,7 +246,7 @@ object Helpers:
       .buildClassInstantiationExpr(ctx, TypeRepr.of[T], idents)
       .asExprOf[T]
 
-    val reqRefExpr = Ref(reqSym).asExprOf[Int]
+    val reqRefExpr = Ref(reqSym).asExprOf[Long]
     val requiredMaskExpr = Expr(requiredMask)
 
     val parseLogic: Term = '{
@@ -266,7 +266,7 @@ object Helpers:
         else
           throw new JsonParseError(
             "Missing required field(s) " + ${ Expr(classRef.fields.map(_.name)) }(
-              Integer.numberOfTrailingZeros($reqRefExpr & $requiredMaskExpr)
+              java.lang.Long.numberOfTrailingZeros($reqRefExpr & $requiredMaskExpr)
             ),
             $in
           )
@@ -290,7 +290,7 @@ object Helpers:
     val (varDefs, idents, reqVarDef, requiredMask, fieldSymbols) =
       FieldDefaultBuilder.generateDefaults[T](ctx, classRef)
     val reqSym = reqVarDef.symbol
-    val reqRefExpr = Ref(reqSym).asExprOf[Int]
+    val reqRefExpr = Ref(reqSym).asExprOf[Long]
     val requiredMaskExpr = Expr(requiredMask)
 
     val instanceSym = Symbol.newVal(Symbol.spliceOwner, "_instance", TypeRepr.of[T], Flags.Mutable, Symbol.noSymbol)
@@ -329,7 +329,7 @@ object Helpers:
       Expr(classRef.fields.map(_.name))
 
     val missingFieldExpr =
-      '{ $ctorFieldNamesExpr(Integer.numberOfTrailingZeros($reqRefExpr & $requiredMaskExpr)) }
+      '{ $ctorFieldNamesExpr(java.lang.Long.numberOfTrailingZeros($reqRefExpr & $requiredMaskExpr)) }
 
     val parseLogic: Term =
       '{
